chore: get local request dvos
